feat(dao): Delete CombinedDAO and port its tests
* Port all needed CombinedDAO tests to JobDAOActor
(as most of the code from CombinedDAO was moved to
 JobDAOActor)
* Delete CombinedDAO as it's not used anymore
