+++ job-server/src/main/resources/application.conf
@@ -32,8 +32,6 @@ spark {
     # - auto - Use SPARK_LOCAL_IP or automatically find externally accessible IPv4
     network-address-resolver = "akka"
 
-    jobdao = spark.jobserver.io.CombinedDAO
-
     dao-timeout = 6s
 
     # At jobserver startup, a cleanup class can be provided which cleans the dao.
