+++ job-server/src/test/resources/local.test.dao.conf
@@ -1,10 +1,5 @@
-# This file contains common test settings for CombinedDAO and ZookeeperDAO
-
 spark.jobserver {
   cache-on-upload = false
-
-  jobdao = spark.jobserver.io.CombinedDAO
-
   combineddao {
     rootdir = /tmp/spark-job-server-test
     binarydao {
