+++ job-server/src/test/resources/local.test.metadatasqldao.conf
@@ -1,8 +1,6 @@
 # This file contains test settings for: MetaDataSqlDAOSpec
 
 spark.jobserver {
-  jobdao = spark.jobserver.io.CombinedDAO
-
   cache-on-upload = true
 
   combineddao {
